callListdbservice updated

+++ server/src/service/callListdb.service.ts
@@ -10,6 +10,9 @@ class CallListDBService implements ICallListService {
     }
 
     async createCallList(title: string, creator: string, description: string): Promise<CallList> {
+        if (! title) {
+            throw new Error("Missing title\n");
+        }
         const cm = await callListModel;
         return await cm.create({
             id : new Date().valueOf(),
@@ -21,6 +24,15 @@ class CallListDBService implements ICallListService {
     }
 
     async editCallList(clId: number, clTitle: string, clCreator: string, clContacts: Array<Contact["id"]>, clDescription: string): Promise<CallList> {
+        if (clId < 0) {
+            throw new Error("Id can't be negative\n");
+        }
+        if (!clTitle) {
+            throw new Error("Missing title\n");
+        }
+        if (!clCreator) {
+            throw new Error("Missing creator\n");
+        }
         const cm = await callListModel;
         await cm.updateOne( {id: clId},
             {title: clTitle, 
@@ -35,6 +47,9 @@ class CallListDBService implements ICallListService {
     }
 
     async deleteCallList(callListId: number): Promise<Boolean> {
+        if (callListId < 0) {
+            throw new Error("Id can't be negative\n");
+        }
         const cm = await callListModel;
         await cm.findByIdAndDelete(callListId);
 
@@ -45,7 +60,6 @@ class CallListDBService implements ICallListService {
         } else return false;
     }
 
-    //måste fixas
     async addContact(callListId: number): Promise<CallList> {
         const contactId = new Date().valueOf();
         const cm = await callListModel;
